+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.2.0
+
+Miscallaneous updates
+
+- throw an error if `-o` argument is set to a directory
+
+- allow user to pass ffmpeg arguments for clip creation
+
+- re-query user if clip start timestamp transcends song length
+
+- not show command stdout/stderr unless invocation fails
+
+- not show `pythmeg: info: invocating command ...` unless invocation fails
+
 ## 1.1.0
 
 Instagram-friendly video output using AAC as output audio codec + usage improvements